Ajout des commentaires pour Doxygen Correction d'un warning dans NeuroneLayer.cpp

+++ sources/neuronelayer.cpp
@@ -1,4 +1,4 @@
-#include "neuronelayer.hpp"
+#include "headers/neuronelayer.hpp"
 
 NeuroneLayer::NeuroneLayer(unsigned int inputSize, unsigned int outputSize, std::function<float(float)> activationF)
 : mPoids(Eigen::MatrixXf::Random(outputSize, inputSize))
@@ -10,7 +10,7 @@ NeuroneLayer::NeuroneLayer(unsigned int inputSize, unsigned int outputSize, std:
 Eigen::VectorXf NeuroneLayer::process(Eigen::VectorXf inputs) const
 {
     Eigen::VectorXf activationLevel = mPoids*inputs;
-    for(size_t i(0); i < activationLevel.size(); i++)
+    for(unsigned int i(0); i < activationLevel.size(); i++)
         activationLevel[i] = mActivationFun(activationLevel[i]);
 
     return activationLevel;
